fix: isolate MCP transport tests
